refactor: InfraStructor 구성
