+++ src/main/java/com/example/surveyapp/domain/order/infrastructor/OrderJpaRepository.java
@@ -1,17 +1,17 @@
-package com.example.surveyapp.domain.order.domain.repository;
+package com.example.surveyapp.domain.order.infrastructor;
 
 import com.example.surveyapp.domain.order.domain.model.Order;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OrderRepository {
+
+public interface OrderJpaRepository extends JpaRepository<Order, Long> {
     Page<Order> findByUserIdAndIsDeletedFalse(Long userId, Pageable pageable);
 
     Optional<Order> findByIdAndIsDeletedFalse(Long id);
 
     Page<Order> findAll(Pageable pageable);
-
-    Order save(Order order);
 }
